requirements.txt
flask
werkzeug
pillow 